refactor(core): move fields to widgets
BREAKING CHANGE: Relocate all *Field components to @dhis2/ui-widgets.
They can be accessed from `@dhis2/ui` using named exports.
