+++ packages/core/src/index.js
@@ -10,7 +10,6 @@ export { ButtonStrip } from './ButtonStrip/ButtonStrip.js'
 export { Card } from './Card/Card.js'
 export { CenteredContent } from './CenteredContent/CenteredContent.js'
 export { Checkbox } from './Checkbox/Checkbox.js'
-export { CheckboxField } from './CheckboxField/CheckboxField.js'
 export { Chip } from './Chip/Chip.js'
 export { CircularLoader } from './CircularLoader/CircularLoader.js'
 export { ComponentCover } from './ComponentCover/ComponentCover.js'
@@ -19,13 +18,11 @@ export { DropdownButton } from './DropdownButton/DropdownButton.js'
 export { Field } from './Field/Field.js'
 export { FieldSet } from './FieldSet/FieldSet.js'
 export { FileInput } from './FileInput/FileInput.js'
-export { FileInputField } from './FileInputField/FileInputField.js'
 export { FileList } from './FileList/FileList.js'
 export { FileListItem } from './FileListItem/FileListItem.js'
 export { FileListPlaceholder } from './FileListPlaceholder/FileListPlaceholder.js'
 export { Help } from './Help/Help.js'
 export { Input } from './Input/Input.js'
-export { InputField } from './InputField/InputField.js'
 export { Label } from './Label/Label.js'
 export { Legend } from './Legend/Legend.js'
 export { LinearLoader } from './LinearLoader/LinearLoader.js'
@@ -38,7 +35,6 @@ export { ModalActions } from './ModalActions/ModalActions.js'
 export { ModalContent } from './ModalContent/ModalContent.js'
 export { ModalTitle } from './ModalTitle/ModalTitle.js'
 export { MultiSelect } from './MultiSelect/MultiSelect.js'
-export { MultiSelectField } from './MultiSelectField/MultiSelectField.js'
 export { MultiSelectOption } from './MultiSelectOption/MultiSelectOption.js'
 export { Node } from './Node/Node.js'
 export { Popover } from './Popover/Popover.js'
@@ -46,7 +42,6 @@ export { Popper } from './Popper/Popper.js'
 export { Radio } from './Radio/Radio.js'
 export { Required } from './Required/Required.js'
 export { SingleSelect } from './SingleSelect/SingleSelect.js'
-export { SingleSelectField } from './SingleSelectField/SingleSelectField.js'
 export { SingleSelectOption } from './SingleSelectOption/SingleSelectOption.js'
 export { StackedTable } from './StackedTable/StackedTable.js'
 export { StackedTableBody } from './StackedTable/StackedTableBody.js'
@@ -58,7 +53,6 @@ export { StackedTableRow } from './StackedTable/StackedTableRow.js'
 export { StackedTableRowHead } from './StackedTable/StackedTableRowHead.js'
 export { SplitButton } from './SplitButton/SplitButton.js'
 export { Switch } from './Switch/Switch.js'
-export { SwitchField } from './SwitchField/SwitchField.js'
 export { Tab } from './Tab/Tab.js'
 export { TabBar } from './TabBar/TabBar.js'
 export { Table } from './Table/Table.js'
@@ -71,7 +65,5 @@ export { TableRow } from './Table/TableRow.js'
 export { TableRowHead } from './Table/TableRowHead.js'
 export { Tag } from './Tag/Tag.js'
 export { TextArea } from './TextArea/TextArea.js'
-export { TextAreaField } from './TextAreaField/TextAreaField.js'
-export { FieldSetField } from './FieldSetField/FieldSetField.js'
 export { Layer } from './Layer/Layer.js'
 export { Tooltip } from './Tooltip/Tooltip.js'
