+++ packages/widgets/src/CheckboxField/CheckboxField.stories.e2e.js
