+++ packages/widgets/src/FieldSetField/FieldSetField.js
@@ -2,9 +2,7 @@ import propTypes from '@dhis2/prop-types'
 import React from 'react'
 
 import { sharedPropTypes } from '@dhis2/ui-constants'
-import { Field } from '../Field/Field.js'
-import { FieldSet } from '../FieldSet/FieldSet.js'
-import { Legend } from '../Legend/Legend.js'
+import { Field, FieldSet, Legend } from '@dhis2/ui-core'
 
 /**
  * @module
