+++ packages/widgets/src/FieldSetField/FieldSetField.stories.e2e.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { storiesOf } from '@storybook/react'
 
-import { Checkbox } from '../index.js'
+import { Checkbox } from '@dhis2/ui-core'
 
 import { FieldSetField } from './FieldSetField.js'
