+++ packages/widgets/src/FileInputField/FileInputField.stories.e2e.js
