+++ packages/widgets/src/FileInputFieldWithList/FileInputFieldWithList.stories.e2e.js
