+++ packages/widgets/src/MultiSelectField/MultiSelectField.stories.js
@@ -1,80 +1,81 @@
 import React from 'react'
 import { storiesOf } from '@storybook/react'
 
-import { SingleSelectOption } from '../index.js'
-import { SingleSelectField } from './SingleSelectField.js'
+import { MultiSelectOption } from '@dhis2/ui-core'
+
+import { MultiSelectField } from './MultiSelectField.js'
 
 const defaultProps = {
     label: 'Default label',
-    selected: '1',
+    selected: ['1'],
     onChange: selected =>
         alert(`Selected changed to: ${JSON.stringify(selected, null, 2)}`),
 }
 
 const options = [
-    <SingleSelectOption key="1" value="1" label="one" />,
-    <SingleSelectOption key="2" value="2" label="two" />,
-    <SingleSelectOption key="3" value="3" label="three" />,
-    <SingleSelectOption key="4" value="4" label="four" />,
-    <SingleSelectOption key="5" value="5" label="five" />,
-    <SingleSelectOption key="6" value="6" label="six" />,
-    <SingleSelectOption key="7" value="7" label="seven" />,
-    <SingleSelectOption key="8" value="8" label="eight" />,
-    <SingleSelectOption key="9" value="9" label="nine" />,
-    <SingleSelectOption key="10" value="10" label="ten" />,
+    <MultiSelectOption key="1" value="1" label="one" />,
+    <MultiSelectOption key="2" value="2" label="two" />,
+    <MultiSelectOption key="3" value="3" label="three" />,
+    <MultiSelectOption key="4" value="4" label="four" />,
+    <MultiSelectOption key="5" value="5" label="five" />,
+    <MultiSelectOption key="6" value="6" label="six" />,
+    <MultiSelectOption key="7" value="7" label="seven" />,
+    <MultiSelectOption key="8" value="8" label="eight" />,
+    <MultiSelectOption key="9" value="9" label="nine" />,
+    <MultiSelectOption key="10" value="10" label="ten" />,
 ]
 
-storiesOf('Components/Core/SingleSelectField', module)
+storiesOf('Components/Widgets/MultiSelectField', module)
     .add('Default', () => (
-        <SingleSelectField {...defaultProps}>{options}</SingleSelectField>
+        <MultiSelectField {...defaultProps}>{options}</MultiSelectField>
     ))
     .add('With Help text', () => (
-        <SingleSelectField {...defaultProps} helpText="A helpful text.">
+        <MultiSelectField {...defaultProps} helpText="A helpful text.">
             {options}
-        </SingleSelectField>
+        </MultiSelectField>
     ))
     .add('Status: Valid', () => (
-        <SingleSelectField
+        <MultiSelectField
             {...defaultProps}
             helpText="A helpful text."
             validationText="Totally valid"
             valid
         >
             {options}
-        </SingleSelectField>
+        </MultiSelectField>
     ))
     .add('Status: Warning', () => (
-        <SingleSelectField
+        <MultiSelectField
             {...defaultProps}
             helpText="A helpful text."
             validationText="Hm, not quite, I warn thee!"
             warning
         >
             {options}
-        </SingleSelectField>
+        </MultiSelectField>
     ))
     .add('Status: Error', () => (
-        <SingleSelectField
+        <MultiSelectField
             {...defaultProps}
             helpText="A helpful text."
             validationText="NO! TOTALLY WRONG!"
             error
         >
             {options}
-        </SingleSelectField>
+        </MultiSelectField>
     ))
     .add('Required', () => (
-        <SingleSelectField {...defaultProps} required>
+        <MultiSelectField {...defaultProps} required>
             {options}
-        </SingleSelectField>
+        </MultiSelectField>
     ))
     .add('Input width', () => (
-        <SingleSelectField
+        <MultiSelectField
             inputWidth="200px"
             {...defaultProps}
             label="A very long label indeed, well at least longer than the input field to show how it looks and works and stuff"
             required
         >
             {options}
-        </SingleSelectField>
+        </MultiSelectField>
     ))
