+++ packages/widgets/src/SingleSelectField/SingleSelectField.js
@@ -1,7 +1,7 @@
 import React from 'react'
 
 import propTypes from '@dhis2/prop-types'
-import { Field, SingleSelect, Box } from '../index.js'
+import { Field, SingleSelect, Box } from '@dhis2/ui-core'
 import { sharedPropTypes } from '@dhis2/ui-constants'
 
 /**
