+++ packages/widgets/src/SwitchField/SwitchField.stories.e2e.js
