+++ packages/widgets/src/index.js
@@ -1,2 +1,12 @@
 export { HeaderBar } from './HeaderBar/HeaderBar.js'
 export { OrganisationUnitTree } from './OrganisationUnitTree/OrganisationUnitTree.js'
+
+/* Fields */
+export { CheckboxField } from './CheckboxField/CheckboxField.js'
+export { FieldSetField } from './FieldSetField/FieldSetField.js'
+export { FileInputField } from './FileInputField/FileInputField.js'
+export { InputField } from './InputField/InputField.js'
+export { MultiSelectField } from './MultiSelectField/MultiSelectField.js'
+export { SingleSelectField } from './SingleSelectField/SingleSelectField.js'
+export { SwitchField } from './SwitchField/SwitchField.js'
+export { TextAreaField } from './TextAreaField/TextAreaField.js'
